Fully support multi-port local_port and remote_port
Powershell returns an array if there's more than one port. Allow Array and properly sort the arrays. Also work around a bug where an array in PS gets turned into a Hash with the count and value keys with convert-tojson. I linked to the stackoverflow that discussed the issue within the code so the reason isn't lost on future readers.

Signed-off-by: Tim Smith <tsmith@chef.io>